Architecture changes in sorting: MergeSort abstract class with inherited concrete sorting classes. .mergeSort method creates instances of these classes and calls it's .sort method.
